% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality_list.R
\name{Causlist}
\alias{Causlist}
\title{R6 class that defines DBNs as causality lists}
\arguments{
\item{size}{Number of timeslices of the DBN}

\item{net}{a dbn object}

\item{caus}{a causlist object}
}
\value{
A new 'causlist' object

a causlist object

a dbn object
}
\description{
Constructor of the 'causlist' class

Translate a DBN into a causality list

This function takes as input a network from a DBN and transforms the 
structure into a causality list if it is a valid DBN. Valid DBNs have only
inter-timeslice edges and only allow variables in t_0 to have parents.

Translate a causality list into a DBN network

This function takes as input a causality list and transforms it into a DBN.
}
\details{
A causality list has a list with causal units, a size representing the
Markovian order of the network and a specific node ordering.
}
\section{Fields}{

\describe{
\item{\code{causal_units}}{List of causal units defining the structure}

\item{\code{size}}{Size of the DBN}

\item{\code{ordering}}{String vector defining the order of the nodes in a timeslice}
}}

