% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_main.R
\name{dummy}
\alias{dummy}
\title{Debug function for C++ code}
\usage{
dummy(ordering, size, n_inds)
}
\arguments{
\item{ordering}{Names of the nodes in the network}

\item{size}{Number of timeslices of the DBN. Markovian order 1 equals size 2, and so on.}

\item{n_ind}{Number of particles used in the algorithm.}
}
\description{
Debug function for C++ code
}
