% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_vl_cpp}
\alias{initialize_vl_cpp}
\title{Create a velocity list and initialize it}
\usage{
initialize_vl_cpp(ordering, size)
}
\arguments{
\item{ordering}{a list with the order of the variables in t_0}

\item{size}{the size of the DBN}
}
\value{
a velocity list
}
\description{
Create a velocity list and initialize it
}
