% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{locate_directions}
\alias{locate_directions}
\title{Find the position of 0's or 1's in a Velocity's causality list}
\usage{
locate_directions(vl, pool, cmp)
}
\arguments{
\item{vl}{the Velocity's causality list}

\item{pool}{the list with the positions}

\item{cmp}{the direction to be searched, either 0 or 1}
}
\value{
a list with the Velocity's new causal list and number of operations
}
\description{
Find the position of 0's or 1's in a Velocity's causality list
}
