% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{Particle}
\alias{Particle}
\title{R6 class that defines a Particle in the PSO algorithm}
\arguments{
\item{ordering}{a vector with the names of the nodes in t_0}

\item{size}{number of timeslices of the DBN}

\item{dt}{dataset to evaluate the fitness of the particle}
}
\value{
A new 'Particle' object

The score of the current position
}
\description{
Constructor of the 'Particle' class

Evaluate the score of the particle's position

Evaluate the score of the particle's position.
Updates the local best if the new one is better.
}
\details{
A particle has a Position, a Velocity and a local best
}
\section{Fields}{

\describe{
\item{\code{ps}}{position of the particle}

\item{\code{cl}}{velocity of the particle}

\item{\code{lb}}{local best score obtained}

\item{\code{lb_ps}}{local best position found}
}}

