% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_ctrl.R
\name{PsoCtrl}
\alias{PsoCtrl}
\title{R6 class that defines the PSO controller}
\arguments{
\item{ordering}{a vector with the names of the nodes in t_0}

\item{size}{number of timeslices of the DBN}

\item{n_inds}{number of particles that the algorithm will simultaneously process}

\item{n_threads}{number of threads used during parallel sections. By default, half of the available cores}

\item{n_it}{maximum number of iterations of the pso algorithm}

\item{in_cte}{parameter that varies the effect of the inertia}

\item{gb_cte}{parameter that varies the effect of the global best}

\item{lb_cte}{parameter that varies the effect of the local best}
}
\value{
A new 'PsoCtrl' object
}
\description{
Constructor of the 'PsoCtrl' class
}
\details{
The controller will encapsulate the particles and run the algorithm
}
\section{Fields}{

\describe{
\item{\code{parts}}{list with all the particles in the algorithm}

\item{\code{cl}}{cluster for the parallel computations}

\item{\code{n_it}}{maximum number of iterations of the pso algorithm}

\item{\code{in_cte}}{parameter that varies the effect of the inertia}

\item{\code{gb_cte}}{parameter that varies the effect of the global best}

\item{\code{lb_cte}}{parameter that varies the effect of the local best}

\item{\code{b_ps}}{best position found}

\item{\code{b_scr}}{best score obtained}
}}

