% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomize_vl_cpp}
\alias{randomize_vl_cpp}
\title{Randomize a velocity with the given probabilities}
\usage{
randomize_vl_cpp(vl, probs, seed)
}
\arguments{
\item{vl}{a velocity list}

\item{probs}{the probabilities of each value in the set {-1,0,1}}

\item{seed}{the seed used for random number generation. Ignored if lesser than 0}
}
\value{
a velocity list with randomized values
}
\description{
Randomize a velocity with the given probabilities
}
