% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_main.R
\name{learn_dbn_structure_pso}
\alias{learn_dbn_structure_pso}
\title{Learn a DBN structure with a PSO approach}
\usage{
learn_dbn_structure_pso(dt, size, n_ind = 30, n_it = 20, score = "bge")
}
\arguments{
\item{dt}{a data.table with the data of the network to be trained. Do not previously fold it with the 'dbnR' package.}

\item{size}{Number of timeslices of the DBN. Markovian order 1 equals size 2, and so on.}

\item{n_ind}{Number of particles used in the algorithm.}

\item{n_it}{Maximum number of iterations that the algorithm can perform.}

\item{score}{The score used to evaluate the structures}
}
\value{
A 'dbn' object with the structure of the best network found
}
\description{
Given a dataset and the desired Markovian order, this function returns a DBN
structure ready to be fitted with the 'dbnR' package.
}
