% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{Velocity}
\alias{Velocity}
\title{R6 class that defines velocities affecting causality lists in the PSO}
\description{
The velocities will be defined as a causality list where each element in
a causal unit is a pair (v, node) with v being either 0, 1 or -1. 0 means 
that arc remained the same, 1 means that arc was added and -1 means that arc 
was deleted.
}
\section{Fields}{

\describe{
\item{\code{abs_op}}{Total number of operations 1 or -1 in the velocity}
}}

